Refactor layers.
